requirements.txt
ultralytics
opencv-python
numpy
Pillow
streamlit